==> notebooks/01_data_prep.ipynb <==
import os
import json
import numpy as np
import pandas as pd
import mediapipe as mp

from tqdm import tqdm
import logging

import matplotlib.pyplot as plt
from mediapipe.framework.formats import landmark_pb2

mp_holistic = mp.solutions.holistic
logger = logging.getLogger(__name__)
logger.setLevel("INFO")

# Constants

connections_type = dict(
    face=mp_holistic.FACEMESH_TESSELATION,
    left_hand=mp_holistic.HAND_CONNECTIONS,
    right_hand=mp_holistic.HAND_CONNECTIONS,
    pose=mp_holistic.POSE_CONNECTIONS
)

IMG_SIZE = (512, 512, 3)

train_dataset_path = "../data/raw/train_landmark_files/"
save_dataset_path = "../data/generated/np_landmarks/"
class_path = "../data/raw/train.csv"
class_to_label_path = "../data/raw/sign_to_prediction_index_map.json"


def get_pts_to_img(data: pd.DataFrame) -> np.ndarray:
    image = np.zeros(IMG_SIZE, np.uint8)

    pts = {
        "face": [],
        "left_hand": [],
        "right_hand": [],
        "pose": []
    }

    for elem in data.itertuples():
        pts[elem.type].append(
            landmark_pb2.NormalizedLandmark(
                x=elem.x,
                y=elem.y,
                z=elem.z,
                visibility=1.0
            )
        )

    for body_part, landmarks in pts.items():
        landmark_subset = landmark_pb2.NormalizedLandmarkList(landmark =  landmarks)

        mp.solutions.drawing_utils.draw_landmarks(
            image,
            landmark_subset,
            connections_type[body_part]
        )
        
    return image

for fn in df.frame.unique():
    image = get_pts_to_img(df.query('frame == @fn'))
    plt.figure()
    plt.imshow(image)

def get_train_dataset(train_dataset_path: str, class_path: str, class_to_label_path: str) -> pd.DataFrame:
    x = []
    for dirpath, dirnames, filenames in os.walk(train_dataset_path):
        if filenames:
            for fl in filenames:
                x.append(os.path.join(dirpath, fl))

    print(f"Found {len(x)} files")
    
    class_name = pd.read_csv(class_path)
    
    with open(class_to_label_path, "r") as fp:
        y_label = json.load(fp)
        y_label = pd.DataFrame.from_dict(y_label, orient="index", columns = ["y_label"])
    
    df = pd.DataFrame(
        x, columns=["filename"]
    )
    
    df = df.set_index("filename").join(class_name.set_index("path")).reset_index()
    df = df.set_index("sign").join(y_label).reset_index()

    return df

def get_frames_count(x):
    fl = pd.read_parquet(x)
    return fl.frame.nunique()

dt = []
for elem in tqdm(df.itertuples(), total = df.shape[0]):
    dt.append(get_frames_count(elem.filename))
    
df["nframes"] = dt

df.to_csv("train_data_with_label.csv")

def get_transformed_data(dataset_files: str, save_dataset_path: str) -> None:
    for fl in tqdm(dataset_files):
        df = pd.read_parquet(fl)
        logger.info(f"Found: {df.frame.nunique()} frames in {fl}")

        all_imgs = []

        for single_frame in df.frame.unique():
            temp = df[df.frame == single_frame]
            single_frame_img = get_pts_to_img(temp)
            all_imgs.append(single_frame_img)

        all_imgs = np.array(all_imgs)

        fl_save = f'{fl.split("/")[-1].split(".parquet")[0]}.npz'

        with open(os.path.join(save_dataset_path, fl_save), 'wb') as fp:
            np.save(fp, all_imgs)
            logger.info(f"Saved {fl_save}")

dataset_files = get_all_files(train_dataset_path)

get_transformed_data(dataset_files=dataset_files, save_dataset_path=save_dataset_path)

Saving as np arrays

all_files = []

for dir, _, fls in os.walk(train_dataset_path):
    for fl in fls:
        all_files.append(os.path.join(dir, fl))

print(len(all_files))

df = pd.read_parquet(all_files[0])

pd.set_option('display.max_rows', None)

all_files[0]

train = pd.read_csv(class_path)
print(train.sequence_id.nunique())
train.groupby("participant_id").path.count()


for fl in tqdm(all_files):
    df = pd.read_parquet(fl)
    sequence_id = train[train.path == fl].sequence_id.iloc[0]

    all_imgs = np.array(
        df.loc[:, ["x", "y", "z"]],
        dtype=np.float16
    )
    all_imgs = all_imgs.reshape((df.frame.nunique(), 543 * 3))

    with open(
        os.path.join("./test", str(sequence_id) + ".npy"), "wb"
    ) as fr:
        np.save(file=fr, arr=all_imgs)
